Ported more code from `stringio.pas` to C.
* stringio.pas: Ported "iCASE", "RotStrL", "RotStrR", "ExpC2StrL", "ExpC2StrR", "ExpC3StrL", "ExpC3StrR", "CenterStr", "CutStrL", "FlipStr", "FilterStr1" and "Bpm2str" routines to C.
* stringio.c: Added "iCASE", "RotStrL", "RotStrR", "ExpC2StrL", "ExpC2StrR", "ExpC3StrL", "ExpC3StrR", "CenterStr", "CutStrL", "FlipStr", "FilterStr1" and "Bpm2str" routines (ported from "stringio.pas").
* stringio.h: Updated due to changes in "stringio.c".
* stringio/i386/FilterStr1.pas: Initial release (saved from "stringio.pas").
* stringio/i386/FlipStr.pas: Initial release (saved from "stringio.pas").
* stringio/i386/iCASE.pas: Initial release (saved from "stringio.pas").
* stringio/pas/Bpm2str.pas: Initial release (saved from "stringio.pas").
* stringio/pas/CenterStr.pas: Initial release (saved from "stringio.pas").
* stringio/pas/CutStrL.pas: Initial release (saved from "stringio.pas").
* stringio/pas/ExpC2StrL.pas: Initial release (saved from "stringio.pas").
* stringio/pas/ExpC2StrR.pas: Initial release (saved from "stringio.pas").
* stringio/pas/ExpC3StrL.pas: Initial release (saved from "stringio.pas").
* stringio/pas/ExpC3StrR.pas: Initial release (saved from "stringio.pas").
* stringio/pas/FilterStr1.pas: Initial release (saved from "stringio.pas").
* stringio/pas/FlipStr.pas: Initial release (saved from "stringio.pas").
* stringio/pas/iCASE.pas: Initial release (saved from "stringio.pas").
* stringio/pas/RotStrL.pas: Initial release (saved from "stringio.pas").
* stringio/pas/RotStrR.pas: Initial release (saved from "stringio.pas").
* stringio/Bpm2str.c: Initial release (ported from "stringio/pas/Bpm2str.pas").
* stringio/CenterStr.c: Initial release (ported from "stringio/pas/CenterStr.pas").
* stringio/CutStrL.c: Initial release (ported from "stringio/pas/CutStrL.pas").
* stringio/ExpC2StrL.c: Initial release (ported from "stringio/pas/ExpC2StrL.pas").
* stringio/ExpC2StrR.c: Initial release (ported from "stringio/pas/ExpC2StrR.pas").
* stringio/ExpC3StrL.c: Initial release (ported from "stringio/pas/ExpC3StrL.pas").
* stringio/ExpC3StrR.c: Initial release (ported from "stringio/pas/ExpC3StrR.pas").
* stringio/FilterStr1.c: Initial release (ported from "stringio/pas/FilterStr1.pas").
* stringio/FlipStr.c: Initial release (ported from "stringio/pas/FlipStr.pas").
* stringio/iCASE.c: Initial release (ported from "stringio/pas/iCASE.pas").
* stringio/RotStrL.c: Initial release (ported from "stringio/pas/RotStrL.pas").
* stringio/RotStrR.c: Initial release (ported from "stringio/pas/RotStrR.pas").
* pascal/string.c: Small fix in "Copy" routine.

+++ pascal/string.c
@@ -121,7 +121,7 @@ String *AppendString (String *dest, const String *src, uint8_t n) {
 
 String_t Copy (const String *s, ssize_t index, ssize_t count) {
   String_t r;
-  int_least8_t len = Length (s);
+  ssize_t len = Length (s);
 
   if (count < 0)
     count = 0;
